+++ src/main/java/roomescape/reservation/service/ThemeService.java
@@ -6,9 +6,9 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import roomescape.reservation.domain.Theme;
 import roomescape.reservation.domain.ThemeName;
-import roomescape.reservation.dto.PopularThemeResponse;
-import roomescape.reservation.dto.ThemeResponse;
-import roomescape.reservation.dto.ThemeSaveRequest;
+import roomescape.reservation.dto.response.PopularThemeResponse;
+import roomescape.reservation.dto.response.ThemeResponse;
+import roomescape.reservation.dto.request.ThemeSaveRequest;
 import roomescape.reservation.repository.ThemeRepository;
 
 @Service
