+++ src/main/java/roomescape/validator/DateRangeValidator.java
@@ -3,7 +3,7 @@ package roomescape.validator;
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
 import java.time.temporal.ChronoUnit;
-import roomescape.reservation.dto.ReservationSearchCondRequest;
+import roomescape.reservation.dto.request.ReservationSearchCondRequest;
 
 public class DateRangeValidator implements ConstraintValidator<ValidDateRange, ReservationSearchCondRequest> {
